pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "karpentermig"
version = "0.2.0"
description = "A tool for Karpenter migration"
readme = "README.md"
authors = [{name = "Pugar Jayanegara"}]
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "questionary",
    "boto3",
    "click",
    "pyyaml",
]

[project.scripts]
karpentermig = "karpentermig.cli:cli"
